requirements.txt
requests
PyYAML
Flask
SQLAlchemy
logging